🔄 Refactor : SoC controller
