+++ back/src/main/java/com/example/mistakes/api/questions/controller/QuestionController.java
@@ -1,9 +1,11 @@
-package com.example.mistakes.api.questions;
+package com.example.mistakes.api.questions.controller;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
+
+import com.example.mistakes.api.questions.QuestionEntity;
 
 @RestController
 @RequestMapping("/api/questions")
