+++ back/src/main/java/com/example/mistakes/api/questions/controller/QuestionResponse.java
@@ -1,5 +1,6 @@
-package com.example.mistakes.api.questions;
+package com.example.mistakes.api.questions.controller;
 
+import com.example.mistakes.api.questions.QuestionEntity;
 import com.example.mistakes.base.type.definition.ResponseMany;
 
 record QuestionResponse(QuestionEntity[] result, int length) implements ResponseMany<Object> {}
