+++ back/src/main/java/com/example/mistakes/api/questions/controller/QuestionResponse2.java
@@ -1,5 +1,6 @@
-package com.example.mistakes.api.questions;
+package com.example.mistakes.api.questions.controller;
 
+import com.example.mistakes.api.questions.QuestionEntity;
 import com.example.mistakes.base.type.template.ResponseManyDTO;
 
 final class QuestionResponse2 extends ResponseManyDTO<QuestionEntity> {
